chore: Upgrade to Yarn v4 and fix publishing
This project has been upgraded from Yarn v1 to Yarn v4. The changes
were based upon the [Yarn migration guide](https://yarnpkg.com/migration/guide)
and the [metamask-module-template](https://github.com/MetaMask/metamask-module-template).

The publishing workflow was also updated because it was impacted by the
Yarn upgrade. The automated publishing should now work correctly, and
should trigger a bot message in Slack when the package is ready for approval.

Additionally, a bunch of dead code/configuration was deleted that was related
to the now-unused CircleCI integration. It was impacted because it also would
have required changes if we kept it.

Closes #197