Carve path to removing reliance on global network
Currently, SwapsController relies on the global network. It uses the
global provider from the NetworkController and uses it to get the
allowance for a user's token on a chain, and it uses the chain ID of the
global network (again, coming from NetworkController) to retrieve
parameters that will ultimately be used to control when and how swap
quotes are requested and how to transform the resulting data. Finally,
it also expects `setProvider` and/or `setChainId` to get called when the
global network is switched, when it will restore previously cached quote
data from state so that it can be presumably reshown in the UI.

Long-term, we want to remove the concept of the global network from the
extension, so we need to prepare this controller accordingly.
Fundamentally, we need to have the parts of the API which rely on the
global provider to take a network client ID, which will be used to fetch
a provider and chain ID.

A couple of things to note about this controller which prohibits us from
completely removing the reliance on the global network:

- Besides `setChainId` and `setProvider`, there is another entry point to
the API via `startFetchAndSetQuotes`. This starts a loop to fetch quotes
on a particular chain. Crucially, this loop does not support multiple
simultaneous network access. If the loop is already active for a network
and `startFetchAndSetQuotes` is called for another, the loop will be
reset and all quotes captured for the previous network will be
overwritten. We plan on keeping this behavior.
- As alluded to above, the controller caches swaps quotes it fetches for
a chain. To retain the previous behavior of restoring from cache when
the network is switched, the controller now listens for the switch via
`NetworkController:networkDidChange`.

Ultimately both of these things may need to be changed when the concept
of the global network is completely removed, but we can revisit this at
a later time.

See the changelog for the full list of changes.