Upgrade Yarn from v1 to 4.5.3
This aligns this repo with our other repos, and provides us with some
nice performance and usability benefits.

This commit updates the GitHub workflow files to match. Constraints also
have been copied over, although they do not match the module template.
This will be addressed in another commit. Finally, with this upgrade we
are able to get rid of an unnecessary entry in `resolutions`.