Upgrade {gas-fee,network}-controller
Upgrade `@metamask/gas-fee-controller` and `@metamask/network-controller`
to their latest versions. This aligns with other upgrades we are making
to controller packages in clients and ensures that we do not run into
peer dependency warnings there when we upgrade
`@metamask/swaps-controller`.

Note that both of these dependencies are actually incorrectly
categorizes as peer dependency when they should just be dependencies.
This will make it even easier to upgrade to a new version of
`@metamask/swaps-controller` in clients. However, that change will be
made in another commit.