Force elliptic to ^6.6.0
There are several security vulnerabilities with the version of
`elliptic` that is present in the dependency tree. We use `elliptic`
indirectly via `@ethersprovider/abi`. We are using the latest version of
this package so we cannot upgrade. Therefore, to use a higher version of
`elliptic` we must force it in `resolutions`.