feat: allow immediate fetching of topAssets without changing the network (#401)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

This PR removes `state.topAssetsLastFetched`, and allows the immediate
fetching of top assets without relying on a
`'NetworkController:networkDidChange'` event. We rely on the
`state.chainCache[chainId].topAssetsLastFetched` instead of the previous
`state.topAssetsLastFetched` now to determine if we can fetch or just
read from the cache.

The use case is for the Mobile Bridge, specifically the situation where
a user is switching back and forth between chain filters but without
actually changing their app network.

`state.topAssetsLastFetched` was never used by Mobile, only in test
mocks for the background state.

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

Related to https://github.com/MetaMask/metamask-mobile/pull/14258

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/swaps-controller`

- **BREAKING**: `state.topAssetsLastFetched` removed
- CHANGED: Rely on `state.chainCache[chainId].topAssetsLastFetched` to
determine cache threshold. This allows for immediate fetching of top
assets from a different chain without actually changing the MM app's
network. Previously if we fetched top assets for Chain A and then
immediately we want to fetch for Chain B, this would be impossible
without actually switching networks. This is needed to fetch top
destination tokens for Bridge transactions.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes